# services/runtime/tests/contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CandidateRuleResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issue_key": {
            "title": "Issue Key",
            "type": "string"
          },
          "knowledge_entry_ids": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Entry Ids",
            "type": "array"
          },
          "recommendation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Recommendation Ids",
            "type": "array"
          },
          "review_ids": {
            "items": {
              "type": "string"
            },
            "title": "Review Ids",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "issue_key",
          "summary",
          "status",
          "recommendation_ids",
          "review_ids",
          "knowledge_entry_ids",
          "created_at"
        ],
        "title": "CandidateRuleResponse",
        "type": "object"
      },
...
      "FeedbackRecordResponse": {
        "properties": {
          "consumed_hint_count": {
            "title": "Consumed Hint Count",
            "type": "integer"
          },
          "consumer_type": {
            "title": "Consumer Type",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "knowledge_entry_ids": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Entry Ids",
            "type": "array"
          },
          "packet_id": {
            "title": "Packet Id",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "subject_key": {
            "title": "Subject Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "packet_id",
          "recommendation_id",
          "consumer_type",
          "subject_key",
          "knowledge_entry_ids",
          "consumed_hint_count",
          "created_at"
        ],
        "title": "FeedbackRecordResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeEntryResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "derived_from": {
            "$ref": "#/components/schemas/KnowledgeRefResponse"
          },
          "evidence_refs": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeRefResponse"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "feedback_targets": {
            "items": {
              "type": "string"
            },
            "title": "Feedback Targets",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "knowledge_type": {
            "title": "Knowledge Type",
            "type": "string"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "narrative",
          "knowledge_type",
          "confidence",
          "feedback_targets",
          "derived_from",
          "evidence_refs",
          "created_at"
        ],
        "title": "KnowledgeEntryResponse",
        "type": "object"
      },
      "KnowledgePacketSummaryResponse": {
        "properties": {
          "governance_hint_count": {
            "title": "Governance Hint Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "intelligence_hint_count": {
            "title": "Intelligence Hint Count",
            "type": "integer"
          },
          "knowledge_entry_ids": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Entry Ids",
            "type": "array"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          }
        },
        "required": [
          "id",
          "recommendation_id",
          "knowledge_entry_ids",
          "governance_hint_count",
          "intelligence_hint_count"
        ],
        "title": "KnowledgePacketSummaryResponse",
        "type": "object"
      },
      "KnowledgeRefResponse": {
        "properties": {
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          }
        },
        "required": [
          "object_type",
          "object_id",
          "relation"
        ],
        "title": "KnowledgeRefResponse",
        "type": "object"
      },
      "KnowledgeRetrieveResponse": {
        "properties": {
          "advisory_only": {
            "default": true,
            "title": "Advisory Only",
            "type": "boolean"
          },
          "candidate_rules": {
            "items": {
              "$ref": "#/components/schemas/CandidateRuleResponse"
            },
            "title": "Candidate Rules",
            "type": "array"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "feedback_records": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FeedbackRecordResponse"
            },
            "title": "Feedback Records",
            "type": "array"
          },
          "packets": {
            "items": {
              "$ref": "#/components/schemas/KnowledgePacketSummaryResponse"
            },
            "title": "Packets",
            "type": "array"
          },
          "recurring_issues": {
            "items": {
              "$ref": "#/components/schemas/RecurringIssueResponse"
            },
            "title": "Recurring Issues",
            "type": "array"
          },
          "root_id": {
            "title": "Root Id",
            "type": "string"
          },
          "root_type": {
            "title": "Root Type",
            "type": "string"
          }
        },
        "required": [
          "root_type",
          "root_id",
          "entries",
          "packets",
          "recurring_issues",
          "candidate_rules"
        ],
        "title": "KnowledgeRetrieveResponse",
        "type": "object"
      },
...
      "RecommendationResponse": {
        "properties": {
          "action_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Summary"
          },
          "adopted": {
            "title": "Adopted",
            "type": "boolean"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "outcome_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Status"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "id",
          "status",
          "created_at",
          "adopted",
          "metadata"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
...
      "RecurringIssueResponse": {
        "properties": {
          "issue_key": {
            "title": "Issue Key",
            "type": "string"
          },
          "knowledge_entry_ids": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Entry Ids",
            "type": "array"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "recommendation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Recommendation Ids",
            "type": "array"
          },
          "review_ids": {
            "items": {
              "type": "string"
            },
            "title": "Review Ids",
            "type": "array"
          },
          "sample_narratives": {
            "items": {
              "type": "string"
            },
            "title": "Sample Narratives",
            "type": "array"
          }
        },
        "required": [
          "issue_key",
          "occurrence_count",
          "sample_narratives",
          "recommendation_ids",
          "review_ids",
          "knowledge_entry_ids"
        ],
        "title": "RecurringIssueResponse",
        "type": "object"
      },
...
      "ReviewDetailResponse": {
        "properties": {
          "cause_tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cause Tags",
            "type": "array"
          },
          "complete_execution_receipt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complete Execution Receipt Id"
          },
          "complete_execution_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complete Execution Request Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "expected_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Outcome"
          },
          "followup_actions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Followup Actions",
            "type": "array"
          },
          "governance_hint_count": {
            "default": 0,
            "title": "Governance Hint Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "intelligence_hint_count": {
            "default": 0,
            "title": "Intelligence Hint Count",
            "type": "integer"
          },
          "knowledge_feedback_packet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Feedback Packet Id"
          },
          "latest_outcome_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Outcome Reason"
          },
          "latest_outcome_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Outcome Status"
          },
          "lessons": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Lessons",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "observed_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Outcome"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          },
          "review_type": {
            "title": "Review Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submit_execution_receipt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submit Execution Receipt Id"
          },
          "submit_execution_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submit Execution Request Id"
          },
          "variance_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variance Summary"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "required": [
          "id",
          "review_type",
          "status",
          "created_at"
        ],
        "title": "ReviewDetailResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AegisOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/recommendations/{recommendation_id}": {
      "get": {
        "operationId": "get_knowledge_for_recommendation_api_v1_knowledge_recommendations__recommendation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeRetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge For Recommendation",
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/v1/knowledge/reviews/{review_id}": {
      "get": {
        "operationId": "get_knowledge_for_review_api_v1_knowledge_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeRetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge For Review",
        "tags": [
          "Knowledge"
        ]
      }
    },
...
    "/api/v1/recommendations/{reco_id}": {
      "get": {
        "operationId": "get_recommendation_detail_api_v1_recommendations__reco_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reco_id",
            "required": true,
            "schema": {
              "title": "Reco Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recommendation Detail",
        "tags": [
          "Recommendations"
        ]
      }
    },
...
    "/api/v1/reviews/{review_id}": {
      "get": {
        "operationId": "get_review_detail_api_v1_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review Detail",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  }
}